requirements.txt
python-telegram-bot~=20.5
sqlalchemy~=1.5
psycopg2-binary~=2.9
redis~=4.5
rq~=1.11
fastapi~=0.95
uvicorn~=0.22
python-dotenv~=1.0
pytest~=7.3
